implement descriptor_callback_func

+++ Structs/USB_DESCRIPTOR_REQUEST.cs
@@ -5,7 +5,7 @@ namespace UsbPcapDotNet;
 [StructLayout(LayoutKind.Sequential)]
 struct USB_DESCRIPTOR_REQUEST
 {
-    public int ConnectionIndex;
+    public ulong ConnectionIndex;
     public USB_SETUP_PACKET SetupPacket;
     /// <summary>
     /// can be USB_STRING_DESCRIPTOR, <see cref="USB_DEVICE_DESCRIPTOR"/> or <see cref="USB_CONFIGURATION_DESCRIPTOR" />
